Remove SourceFile wrapper
